+++ iotea/src/main/java/hu/zza/iotea/model/response/Problem.java
@@ -6,18 +6,28 @@ import org.zalando.problem.*;
 
 @Getter
 public abstract class Problem extends AbstractThrowableProblem implements Response {
-  private final Long deviceId;
-  private final Long commandId;
+  private final Integer deviceId;
+  private final Integer commandId;
 
   protected Problem(
-      Long deviceId, Long commandId, URI type, String title, StatusType status, String detail) {
+      Integer deviceId,
+      Integer commandId,
+      URI type,
+      String title,
+      StatusType status,
+      String detail) {
     super(type, title, status, detail);
     this.deviceId = deviceId;
     this.commandId = commandId;
   }
 
   protected Problem(
-      Long deviceId, Long commandId, String type, String title, int statusCode, String detail) {
+      Integer deviceId,
+      Integer commandId,
+      String type,
+      String title,
+      int statusCode,
+      String detail) {
     this(deviceId, commandId, URI.create(type), title, Status.valueOf(statusCode), detail);
   }
