+++ iotea/src/main/java/hu/zza/iotea/model/response/ServiceProblem.java
@@ -3,7 +3,7 @@ package hu.zza.iotea.model.response;
 import static java.net.HttpURLConnection.HTTP_BAD_REQUEST;
 
 public class ServiceProblem extends Problem {
-  public ServiceProblem(Long deviceId, Long commandId, String detail) {
+  public ServiceProblem(Integer deviceId, Integer commandId, String detail) {
     super(
         deviceId,
         commandId,
